Added router link to register page
